исправил проблемы объединения

+++ Models/Services/FilesService.cs
@@ -1,8 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 using Models.DatabaseEntities;
-using Models.DTO.Files;
 using Models.Extensions;
 using Models.Interfaces.Repository;
 using Models.Models;
@@ -14,9 +14,15 @@ namespace Models.Services
 
         private readonly IFilesRepository _filesRepository;
 
-        public FilesService(IFilesRepository filesRepository)
+        private readonly IGlossaryRepository _glossaryRepository;
+
+        public FilesService(
+            IFilesRepository filesRepository,
+            IGlossaryRepository glossaryRepository
+            )
         {
             this._filesRepository = filesRepository;
+            this._glossaryRepository = glossaryRepository;
         }
 
         public async Task<IEnumerable<Node<File>>> GetAll()
@@ -25,10 +31,19 @@ namespace Models.Services
             return files?.ToTree((file, icon) => new Node<File>(file, icon), (file) => GetIconByFile(file));
         }
 
-        public async Task<IEnumerable<Node<File>>> GetByProjectIdAsync(int projectId)
+        public async Task<IEnumerable<Node<File>>> GetByProjectIdAsync(int projectId, string fileNamesSearch)
         {
-            var files = await this._filesRepository.GetByProjectIdAsync(projectId: projectId);
-            return files?.ToTree((file, icon) => new Node<File>(file, icon), (file) => GetIconByFile(file));
+            IEnumerable<File> files;
+            if (string.IsNullOrEmpty(fileNamesSearch))
+            {
+                files = await this._filesRepository.GetByProjectIdAsync(projectId: projectId);
+                return files?.ToTree((file, icon) => new Node<File>(file, icon), (file) => this.GetIconByFile(file));
+            }
+            else
+            {
+                files = await this._filesRepository.GetByProjectIdAsync(projectId: projectId, fileNamesSearch: fileNamesSearch);
+                return files?.Select(file => new Node<File>(file, this.GetIconByFile(file)));
+            }
         }
 
         public async Task<File> GetById(int id)
@@ -71,7 +86,7 @@ namespace Models.Services
                 ID_LocalizationProject = projectId,
             };
 
-            return await this.AddNode(newFile);
+            return await this.AddNode(newFile, insertToDbAction: this.InsertFileToDbAsync);
         }
 
         public async Task<Node<File>> AddFolder(FolderModel newFolderModel)
@@ -89,7 +104,7 @@ namespace Models.Services
                 ID_FolderOwner = newFolderModel.ParentId,
                 ID_LocalizationProject = newFolderModel.ProjectId
             };
-            return await AddNode(newFolder);
+            return await AddNode(newFolder, insertToDbAction: this.InsertFolderToDbAsync);
         }
 
         public async Task UpdateNode(int id, File file)
@@ -98,7 +113,7 @@ namespace Models.Services
             // var foundedFile = await filesRepository.GetByID(id);
             // if (foundedFile == null)
             // {
-            //     return NotFound($"File by id \"{id}\" not found");
+            //     throw new Exception($"File by id \"{id}\" not found");
             // }
 
             file.ID = id;
@@ -116,8 +131,14 @@ namespace Models.Services
             // var foundedFile = await filesRepository.GetByID(id);
             // if (foundedFile == null)
             // {
-            //     return NotFound($"File by id \"{id}\" not found");
+            //     throw new Exception($"File by id \"{id}\" not found");
             // }
+
+            var glossary = await this._glossaryRepository.GetByFileIdAsync(id);
+            if (glossary != null)
+            {
+                throw new Exception("Deletion of glossary file is forbidden.");
+            }
 
             var deleteSuccessfully = await this._filesRepository.RemoveAsync(id);
             if (!deleteSuccessfully)
@@ -126,26 +147,40 @@ namespace Models.Services
             }
         }
 
-        private async Task<Node<File>> AddNode(File file)
+        private async Task<Node<File>> AddNode(File file, Func<File, Task> insertToDbAction)
         {
             if (file.ID_FolderOwner.HasValue)
             {
                 var parentFile = await this._filesRepository.GetByIDAsync(file.ID_FolderOwner.Value);
                 if (parentFile?.IsFolder == false)
                 {
-                    throw new Exception($"Can not add new node \"{file.Name}\" in file node");
+                    throw new Exception($"Can not add new node \"{file.Name}\" with file as parent node");
                 }
             }
 
-            var fileUploaded = await this._filesRepository.Upload(file);
-            if (!fileUploaded)
-            {
-                throw new Exception($"Failed to insert folder/file \"{file.Name}\" in database");
-            }
+            await insertToDbAction(file);
 
             var addedFile = await this._filesRepository.GetByNameAndParentId(file.Name, file.ID_FolderOwner);
             var icon = GetIconByFile(addedFile);
             return new Node<File>(addedFile, icon);
+        }
+
+        private async Task InsertFileToDbAsync(File file)
+        {
+            var fileUploaded = await this._filesRepository.Upload(file);
+            if (!fileUploaded)
+            {
+                throw new Exception($"Failed to insert file \"{file.Name}\" in database");
+            }
+        }
+
+        private async Task InsertFolderToDbAsync(File file)
+        {
+            var folderAdded = await this._filesRepository.AddAsync(file) > 0;
+            if (!folderAdded)
+            {
+                throw new Exception($"Failed to insert folder \"{file.Name}\" in database");
+            }
         }
 
         private string GetIconByFile(File file)
